rename heatmap to activity
