+++ applications/fishing-map/src/features/analysis/AnalysisLayerPanel.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import Tooltip from '@globalfishingwatch/ui-components/dist/tooltip'
 import { DatasetTypes } from '@globalfishingwatch/api-types'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
-import { isFishingDataview, isPresenceDataview } from 'features/workspace/heatmaps/heatmaps.utils'
+import { isFishingDataview, isPresenceDataview } from 'features/workspace/activity/activity.utils'
 import DatasetSchemaField from 'features/workspace/shared/DatasetSchemaField'
 import DatasetFilterSource from 'features/workspace/shared/DatasetSourceField'
 import DatasetFlagField from 'features/workspace/shared/DatasetFlagsField'
