+++ applications/fishing-map/src/features/feedback/FeedbackModal.tsx
@@ -5,7 +5,7 @@ import Modal from '@globalfishingwatch/ui-components/dist/modal'
 import { Button, InputText, Select } from '@globalfishingwatch/ui-components'
 import { Generators } from '@globalfishingwatch/layer-composer'
 import { selectActiveDataviews } from 'features/dataviews/dataviews.selectors'
-import { getSourcesSelectedInDataview } from 'features/workspace/heatmaps/heatmaps.utils'
+import { getSourcesSelectedInDataview } from 'features/workspace/activity/activity.utils'
 import { GUEST_USER_TYPE, selectUserData } from 'features/user/user.slice'
 import { loadSpreadsheetDoc } from 'utils/spreadsheet'
 import { isGuestUser, selectUserGroupsClean } from 'features/user/user.selectors'
