+++ applications/fishing-map/src/features/workspace/activity/ActivityFilters.module.css
