+++ applications/fishing-map/src/features/workspace/activity/ActivitySection.module.css
