+++ applications/fishing-map/src/features/workspace/activity/activity.utils.ts
